feat: (임시) root 경로일 때 home으로 리다이렉트
